fix to sql schema and onlinestore code

+++ sql/onlinestore_schema.sql
@@ -1,6 +1,6 @@
 CREATE DATABASE onlinestore;
 CREATE TABLE onlinestore.products (
-    product_id int NOT NULL,
+    product_id bigint NOT NULL,
     product_name varchar(255) NOT NULL,
     product_description varchar(255),
     price DECIMAL(10,2) NOT NULL,
